stub in dynamic views
