+++ android-example/app/src/main/res/layout/fragment_employee.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SecondFragment">
+    tools:context=".EmployeeFragment">
 
     <TextView
         android:id="@+id/textview_second"
