+++ android-example/app/src/main/res/values/strings.xml
@@ -2,11 +2,14 @@
     <string name="app_name">Acme GB</string>
     <string name="action_settings">Settings</string>
     <!-- Strings used for fragments for navigation -->
-    <string name="first_fragment_label">First Fragment</string>
-    <string name="second_fragment_label">Second Fragment</string>
+    <string name="fragment_public_title">Acme Donuts (Public)</string>
+    <string name="fragment_employee_title">Donut Claim Centre</string>
+
+    <string name="donut_price_message">Your price for donuts is $%1$s</string>
+
     <string name="next">Next</string>
     <string name="previous">Previous</string>
 
-    <string name="hello_first_fragment">Hello first fragment</string>
+    <string name="fragment_public_welcome_text">Acme Donuts app</string>
     <string name="hello_second_fragment">Hello second fragment. Arg: %1$s</string>
 </resources>
